Remove OpenMPOp, inlining its remainder into Elemwise
Elemwise was the only consumer. It now derives from COp directly and carries
the openmp request plus _use_openmp/c_compile_args itself; the dead omp.h
header path (always shadowed by Elemwise.c_headers) is dropped.
